+++ dubbo-xds/src/main/java/org/apache/dubbo/xds/PilotExchanger.java
@@ -17,8 +17,6 @@
 package org.apache.dubbo.xds;
 
 import org.apache.dubbo.common.URL;
-import org.apache.dubbo.xds.directory.XdsDirectory;
-import org.apache.dubbo.xds.directory.XdsResourceListener;
 import org.apache.dubbo.xds.resource.XdsResourceType;
 import org.apache.dubbo.xds.resource.update.ResourceUpdate;
 
@@ -28,8 +26,8 @@ public class PilotExchanger {
 
     private static PilotExchanger GLOBAL_PILOT_EXCHANGER = null;
 
-    protected PilotExchanger(URL url) {
-        adsObserver = new AdsObserver(url);
+    protected PilotExchanger() {
+        adsObserver = new AdsObserver();
     }
 
     public <T extends ResourceUpdate> void subscribeXdsResource(
@@ -41,29 +39,13 @@ public class PilotExchanger {
         adsObserver.addListener(resourceName, resourceType, resourceListener);
     }
 
-    public void unSubscribeXdsResource(String clusterName, XdsDirectory listener) {}
-
-    public static PilotExchanger initialize(URL url) {
+    public static PilotExchanger getInstance() {
         synchronized (PilotExchanger.class) {
             if (GLOBAL_PILOT_EXCHANGER != null) {
                 return GLOBAL_PILOT_EXCHANGER;
             }
-            return (GLOBAL_PILOT_EXCHANGER = new PilotExchanger(url));
+            return (GLOBAL_PILOT_EXCHANGER = new PilotExchanger());
         }
-    }
-
-    public static PilotExchanger getInstance() {
-        synchronized (PilotExchanger.class) {
-            return GLOBAL_PILOT_EXCHANGER;
-        }
-    }
-
-    public static PilotExchanger createInstance(URL url) {
-        return new PilotExchanger(url);
-    }
-
-    public static boolean isEnabled() {
-        return GLOBAL_PILOT_EXCHANGER != null;
     }
 
     public void destroy() {
