--- dubbo-xds/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.registry.client.ServiceDiscoveryFactory
@@ -1,1 +1,0 @@
-xds=org.apache.dubbo.xds.registry.XdsServiceDiscoveryFactory
